Remove float to int warnings

+++ ColorMagic/manipulation/color_distance.cpp
@@ -108,9 +108,9 @@ float color_manipulation::color_distance::cielab_delta_e_cie00(color_space::colo
 	auto temp_avg_c = (temp_c1 + temp_c2) / 2.f;
 
 	auto h1 = (int)(to_deg(atan2f(color1_lab->b(), temp_a1))) % 360;
-	if (h1 < 0.f) h1 += 360.f;
+	if (h1 < 0.f) h1 += 360;
 	auto h2 = (int)(to_deg(atan2f(color2_lab->b(), temp_a2))) % 360;
-	if (h2 < 0.f) h2 += 360.f;
+	if (h2 < 0.f) h2 += 360;
 
 	float H;
 	if (abs(h1 - h2) > 180.f)
@@ -122,18 +122,18 @@ float color_manipulation::color_distance::cielab_delta_e_cie00(color_space::colo
 		H = (h1 + h2) / 2.f;
 	}
 
-	float delta_h;
+	int delta_h;
 	if (abs(h2 - h1) <= 180.f)
 	{
 		delta_h = h2 - h1;
 	}
 	else if (abs(h2 - h1) > 180.f && h2 <= h1)
 	{
-		delta_h = h2 - h1 + 360.f;
+		delta_h = h2 - h1 + 360;
 	}
 	else
 	{
-		delta_h = h2 - h1 - 360.f;
+		delta_h = h2 - h1 - 360;
 	}
 
 	auto delta_L = color2_lab->luminance() - color1_lab->luminance();
